Fix overflow in MCE direct conv cycle estimation
When estimating the number of cycles needed to complete a MCE direct
convolution, several 32-bit unsigned integers are multiplied and the
result is stored in a 32-bit unsigned integer which may overflow.

To resolve this, the result will now be stored in a 64-bit unsigned
integer and a 64-bit multiplication is performed instead.

+++ driver/support_library/src/cascading/MceEstimationUtils.cpp
@@ -94,7 +94,7 @@ uint64_t GetMceCycleCountDirect(const HardwareCapabilities& caps,
     uint32_t w        = utils::RoundUpToNearestMultiple(outputShape[2], halfPatchWidth);
     uint32_t i        = utils::RoundUpToNearestMultiple(ifmChannelsPerOfm, ifmChannelsPerMacUnit);
     uint32_t o        = utils::RoundUpToNearestMultiple(outputShape[3], numActiveOgs);
-    uint32_t macCount = numKernelElements * h * w * i * o;
+    uint64_t macCount = static_cast<uint64_t>(numKernelElements) * h * w * i * o;
 
     uint32_t macsPerCycle = ifmChannelsPerMacUnit * numMacUnitsPerOg * numActiveOgs;
